+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+## [1.0.1] - 2022-10-30
+
+### Added
+
+- Added MIT license ([`b17cb1f`](https://github.com/Level/level/commit/b17cb1f)).
 
 ## [1.0.0] - 2022-10-30
 
 *First Public Release*
 
-[Unreleased]: https://github.com/lmd-code/phpmdcompiler/compare/v1.0.0...HEAD
+[1.0.1]: https://github.com/lmd-code/phpmdcompiler/releases/tag/v1.0.1
 [1.0.0]: https://github.com/lmd-code/phpmdcompiler/releases/tag/v1.0.0